requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pyyaml>=6.0
